preform some type casts to prevent warnings

+++ src/editor_layer.cpp
@@ -61,7 +61,7 @@ namespace GJGO
         // Overlay
         ImGuiWindowFlags overlayFlags = ImGuiWindowFlags_NoDecoration | ImGuiWindowFlags_NoDocking | ImGuiWindowFlags_AlwaysAutoResize | ImGuiWindowFlags_NoSavedSettings |
                                             ImGuiWindowFlags_NoFocusOnAppearing | ImGuiWindowFlags_NoNav | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoBackground | ImGuiWindowFlags_NoInputs;
-        ImGui::SetNextWindowPos({GJGO::g_appInstancePtr->window.width - ImGui::CalcTextSize("16.667 ms/frame (60.0 FPS)").x - 10, 10.0f});
+        ImGui::SetNextWindowPos({static_cast<float>(GJGO::g_appInstancePtr->window.width) - ImGui::CalcTextSize("16.667 ms/frame (60.0 FPS)").x - 10.0f, 10.0f});
         if (ImGui::Begin("Example: Simple overlay", NULL, overlayFlags))
         {
             ImGui::Text("%.3f ms/frame (%.1f FPS)", 1000.0f / io.Framerate, io.Framerate);
